+++ simple/library/changelog.md
@@ -1,5 +1,20 @@
 # WebMan WordPress Theme Framework (Simple) Changelog
 
+## 1.0.13
+
+* **Update**: Renaming the `include` folder to `includes`
+* **Update**: Updating `{%= prefix_constant %}_INCLUDE_DIR` constant name to `{%= prefix_constant %}_INCLUDES_DIR`
+* **Update**: Library files paths
+* **Update**: Theme files paths
+
+#### Files changed:
+
+	init.php
+	includes/customize.php
+	includes/visual-editor.php
+	includes/classes/class-customize.php
+
+
 ## 1.0.12
 
 * **Update**: Compatibility with Readability.com
@@ -7,7 +22,7 @@
 #### Files changed:
 
 	init.php
-	include/classes/class-core.php
+	includes/classes/class-core.php
 
 
 ## 1.0.11
@@ -17,7 +32,7 @@
 #### Files changed:
 
 	init.php
-	include/customize.php
+	includes/customize.php
 
 
 ## 1.0.9
@@ -27,7 +42,7 @@
 #### Files changed:
 
 	init.php
-	include/classes/class-core.php
+	includes/classes/class-core.php
 
 
 ## 1.0.8
@@ -40,9 +55,9 @@
 #### Files changed:
 
 	init.php
-	include/customize.php
-	include/visual-editor.php
-	include/classes/class-customize.php
+	includes/customize.php
+	includes/visual-editor.php
+	includes/classes/class-customize.php
 
 
 ## 1.0.7